% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n_calc.R
\name{n_calc}
\alias{n_calc}
\title{Sample size calculations for superiority, non-inferiority and equivalence trials}
\usage{
n_calc(type, r, alpha, beta, sd, mdiff, lim, drop = 0, known = TRUE)
}
\arguments{
\item{type}{Character with the type of calculation. 'sup', 'equi' or 'ninf'.}

\item{r}{Numeric with the ratio between groups A and B.}

\item{alpha}{Numeric with the alpha.}

\item{beta}{Numeric with the beta.}

\item{sd}{Numeric with the standard deviation.}

\item{mdiff}{Numeric with the differences between means.}

\item{lim}{Numeric with the limit to be considered.}

\item{drop}{Numeric with the dropout rate.}

\item{known}{Boolean if calculation should supose known variance.}
}
\value{
Array with sample size.
}
\description{
Sample size calculations for superiority, non-inferiority and equivalence trials
}
\details{
Flight L, Julious SA. Practical guide to sample size calculations: superiority trials. Pharmaceutical Statistics 2015

Flight L, Julious SA. Practical guide to sample size calculations: non-inferiority and equivalence trials. Pharmaceutical Statistics 2015
}
\examples{
x <- n_calc('ninf', 2, 0.05, 0.1, 0.5, 0, 1, 0, FALSE)
}
