% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poder_calc.R
\name{poder_calc}
\alias{poder_calc}
\title{Power calculations for superiority, non-inferiority and equivalence trials}
\usage{
poder_calc(r, alpha, na, mdiff, sd, lim, known, type)
}
\arguments{
\item{r}{Numeric with the ratio between groups A and B.}

\item{alpha}{Numeric with the alpha.}

\item{na}{Numeric with the sample size for group A.}

\item{mdiff}{Numeric with the differences between means.}

\item{sd}{Numeric with the standard deviation.}

\item{lim}{Numeric with the limit to be considered.}

\item{known}{Boolean if calculation should supose known variance.}

\item{type}{Character with the type of calculation. 'sup', 'equi' or 'ninf'.}
}
\value{
Array with power.
}
\description{
Power calculations for superiority, non-inferiority and equivalence trials
}
\details{
Flight L, Julious SA. Practical guide to sample size calculations: superiority trials. Pharmaceutical Statistics 2015

Flight L, Julious SA. Practical guide to sample size calculations: non-inferiority and equivalence trials. Pharmaceutical Statistics 2015
}
\examples{
x <- n_calc('ninf', 2, 0.05, 0.1, 0.5, 0, 1, 0, FALSE)
}
